Move user-facing aggregator classes out of runners namespace
