+++ sdks/java/core/src/main/java/org/apache/beam/sdk/AggregatorRetrievalException.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.runners;
+package org.apache.beam.sdk;
 
 import org.apache.beam.sdk.transforms.Aggregator;
